chore: 메인페이지 랭킹 섹션 컴포넌트 분할
