+++ src/pages/Main.css
@@ -56,86 +56,6 @@
     font-size: 18px;
 }
 
-.rankingContainer {
-    margin: 120px calc(8% + 120px);
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-}
-
-.rankingTitle {
-    font-size: 35px;
-    font-weight: bold;
-
-}
-
-.rankingSelectWhoContainer {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    margin: 20px;
-}
-
-.rankingWhoItemContainer {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    font-size: 20px;
-    margin: 0 95px;
-    cursor: pointer;
-}
-
-.whoItemImg {
-    width: 60px;
-    height: 60px;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    background-color: #E3EFFE;
-    border-radius: 25px;
-    font-weight: bold;
-    color: #fff;
-    margin: 10px;
-}
-
-.whoItemText {
-    color: #6C6C6C;
-}
-
-.selectedItemWho .whoItemImg {
-    background-color: #719CE9;
-}
-
-.selectedItemWho .whoItemText {
-    color: #4078DF;
-    font-weight: bold;
-}
-
-.rankingSelectBox {
-    display: flex;
-    background-color: #E3EFFE;
-    width: 100%;
-    align-items: center;
-    justify-content: center;
-    border-radius: 10px;
-    border: 1px solid #D9E5FB;
-    margin: 10px;
-}
-
-.rankingBoxItem {
-    margin: 20px 30px;
-    color: #719CE9;
-    font-size: 22px;
-    cursor: pointer;
-}
-
-.selectedItemWhich {
-    color: #4078DF;
-    font-weight: bold;
-}
-
 
 @media (max-width: 992px) {
     .selectFriend {
